feat: implement test for documentation section in graphiQL page
